Fix breakout title length issues. Fixes #134.
Because the slug field had a max length of 50, when the breakout title
was between 50 and 80 characters, there was no error message displayed
(since 80 character titles are OK) but title setting would fail (because
the slug would be too long).

Increase the slug field to 90 chars so that it can include the slugified
title with ID appended.

Fix inconsistency between maximum lengths in in-place-editor and modal
editor.

Fix overflow wrap for title display in breakout list.